+++ front/src/components/pages/Blog/BlogPresenter.tsx
