+++ front/src/components/pages/Blog/index.ts
